Use PolylineCollection for time-dynamic polylines
In order to keep feature parity between static and time-dynamic polylines,
PolylineGeometryUpdater was using PolylineGeometry to render lines that
were changing every frame.  Obviously this generated a lot of garbage and
would kill performance when using a lot of time-dynamic lines (like the d3
demo does).  In order to improve performance, I've updated the dynamic
lines to use PolylineCollection instead; which greatly improves performance.

The only feature PolylineGeometry has that PolylineCollection lacks is
per-vertex colors; but since the DataSource layer doesn't expose that feature
yet; this is okay for now.  Of course when we have dynamic buffers,
this problem should just go away.