Minor PrimitivePipeline cleanup.
`PrimitivePipeline.unpackCombineGeometryParameters` was not including the
cloned `modelMatrix` parameter as part of the result.  Because of this
`PrimitivePipeline.combineGeometry` was unecessarilly creating a dummy
object in order to clone it.  This is no longer needed.

There was also an uneeded clone of the resulting modelMatrix in Primtive.js,
the result matrix is the same instance that is passed into combineGeometry
so no clone is needed.

This change is basically a no-op but makes the code less confusing.