Remove faulty logic from GlobeSurfaceTileProvider when layer removed.
The property is named `renderable`, not `isRenderable`, but it doesn't matter anyway because this code is wrong.  If the property name is corrected, then removing the base layer produces a missing globe instead of the blue globe that you get when you don't add a base layer at construction time.